Refactor imports map to separate util
